refactor(exporter) Shopware mapper rename
