+++ module/exporter-shopware-6/src/Infrastructure/Mapper/Product/AbstractProductCustomFieldSetMapper.php
@@ -14,30 +14,23 @@ use Ergonode\Core\Domain\ValueObject\Language;
 use Ergonode\Exporter\Domain\Entity\Export;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
 use Ergonode\ExporterShopware6\Infrastructure\Calculator\AttributeTranslationInheritanceCalculator;
-use Ergonode\ExporterShopware6\Infrastructure\Client\Shopware6PropertyGroupOptionClient;
-use Ergonode\ExporterShopware6\Infrastructure\Mapper\Shopware6ProductMapperInterface;
+use Ergonode\ExporterShopware6\Infrastructure\Mapper\ProductMapperInterface;
 use Ergonode\ExporterShopware6\Infrastructure\Model\Shopware6Product;
-use Ergonode\ExporterShopware6\Infrastructure\Model\Shopware6PropertyGroupOption;
 use Ergonode\Product\Domain\Entity\AbstractProduct;
 use Ergonode\SharedKernel\Domain\Aggregate\AttributeId;
-use Ergonode\Value\Domain\ValueObject\ValueInterface;
 use Webmozart\Assert\Assert;
 
-abstract class AbstractShopware6ProductPropertyGroupMapper implements Shopware6ProductMapperInterface
+abstract class AbstractProductCustomFieldSetMapper implements ProductMapperInterface
 {
-    protected AttributeRepositoryInterface $repository;
+    private AttributeRepositoryInterface $repository;
 
-    protected Shopware6PropertyGroupOptionClient $propertyGroupOptionClient;
-
-    protected AttributeTranslationInheritanceCalculator $calculator;
+    private AttributeTranslationInheritanceCalculator $calculator;
 
     public function __construct(
         AttributeRepositoryInterface $repository,
-        Shopware6PropertyGroupOptionClient $propertyGroupOptionClient,
         AttributeTranslationInheritanceCalculator $calculator
     ) {
         $this->repository = $repository;
-        $this->propertyGroupOptionClient = $propertyGroupOptionClient;
         $this->calculator = $calculator;
     }
 
@@ -51,7 +44,8 @@ abstract class AbstractShopware6ProductPropertyGroupMapper implements Shopware6P
         AbstractProduct $product,
         ?Language $language = null
     ): Shopware6Product {
-        foreach ($channel->getPropertyGroup() as $attributeId) {
+
+        foreach ($channel->getCustomField() as $attributeId) {
             $this->attributeMap($shopware6Product, $attributeId, $product, $channel, $language);
         }
 
@@ -65,32 +59,16 @@ abstract class AbstractShopware6ProductPropertyGroupMapper implements Shopware6P
         return $this->getType() === $type;
     }
 
-    protected function createPropertyGroupOptions(
+    /**
+     * @param mixed $calculateValue
+     *
+     * @return string|array
+     */
+    abstract protected function getValue(
         Shopware6Channel $channel,
         AbstractAttribute $attribute,
-        ValueInterface $value
-    ): Shopware6PropertyGroupOption {
-        $name = $this->calculator->calculate($attribute, $value, $channel->getDefaultLanguage());
-
-        $propertyGroupOption = new Shopware6PropertyGroupOption(null, $name);
-
-        foreach ($channel->getLanguages() as $language) {
-            $calculateValue = $this->calculator->calculate($attribute, $value, $language);
-            if ($calculateValue) {
-                $propertyGroupOption->addTranslations($language, 'name', $calculateValue);
-            }
-        }
-
-        return $propertyGroupOption;
-    }
-
-    abstract protected function addProperty(
-        Shopware6Product $shopware6Product,
-        AbstractAttribute $attribute,
-        AbstractProduct $product,
-        Shopware6Channel $channel,
-        ?Language $language = null
-    ): Shopware6Product;
+        $calculateValue
+    );
 
     private function attributeMap(
         Shopware6Product $shopware6Product,
@@ -107,7 +85,19 @@ abstract class AbstractShopware6ProductPropertyGroupMapper implements Shopware6P
         }
 
         if ($this->isSupported($attribute->getType())) {
-            $this->addProperty($shopware6Product, $attribute, $product, $channel, $language);
+            $value = $product->getAttribute($attribute->getCode());
+            $calculateValue = $this->calculator->calculate(
+                $attribute,
+                $value,
+                $language ?: $channel->getDefaultLanguage()
+            );
+
+            if ($calculateValue) {
+                $shopware6Product->addCustomField(
+                    $attribute->getCode()->getValue(),
+                    $this->getValue($channel, $attribute, $calculateValue)
+                );
+            }
         }
 
         return $shopware6Product;
