+++ module/exporter-shopware-6/src/Infrastructure/Mapper/Product/AbstractProductPropertyGroupMapper.php
@@ -14,23 +14,30 @@ use Ergonode\Core\Domain\ValueObject\Language;
 use Ergonode\Exporter\Domain\Entity\Export;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
 use Ergonode\ExporterShopware6\Infrastructure\Calculator\AttributeTranslationInheritanceCalculator;
-use Ergonode\ExporterShopware6\Infrastructure\Mapper\Shopware6ProductMapperInterface;
+use Ergonode\ExporterShopware6\Infrastructure\Client\Shopware6PropertyGroupOptionClient;
+use Ergonode\ExporterShopware6\Infrastructure\Mapper\ProductMapperInterface;
 use Ergonode\ExporterShopware6\Infrastructure\Model\Shopware6Product;
+use Ergonode\ExporterShopware6\Infrastructure\Model\Shopware6PropertyGroupOption;
 use Ergonode\Product\Domain\Entity\AbstractProduct;
 use Ergonode\SharedKernel\Domain\Aggregate\AttributeId;
+use Ergonode\Value\Domain\ValueObject\ValueInterface;
 use Webmozart\Assert\Assert;
 
-abstract class AbstractShopware6ProductCustomFieldSetMapper implements Shopware6ProductMapperInterface
+abstract class AbstractProductPropertyGroupMapper implements ProductMapperInterface
 {
-    private AttributeRepositoryInterface $repository;
+    protected AttributeRepositoryInterface $repository;
 
-    private AttributeTranslationInheritanceCalculator $calculator;
+    protected Shopware6PropertyGroupOptionClient $propertyGroupOptionClient;
+
+    protected AttributeTranslationInheritanceCalculator $calculator;
 
     public function __construct(
         AttributeRepositoryInterface $repository,
+        Shopware6PropertyGroupOptionClient $propertyGroupOptionClient,
         AttributeTranslationInheritanceCalculator $calculator
     ) {
         $this->repository = $repository;
+        $this->propertyGroupOptionClient = $propertyGroupOptionClient;
         $this->calculator = $calculator;
     }
 
@@ -44,8 +51,7 @@ abstract class AbstractShopware6ProductCustomFieldSetMapper implements Shopware6
         AbstractProduct $product,
         ?Language $language = null
     ): Shopware6Product {
-
-        foreach ($channel->getCustomField() as $attributeId) {
+        foreach ($channel->getPropertyGroup() as $attributeId) {
             $this->attributeMap($shopware6Product, $attributeId, $product, $channel, $language);
         }
 
@@ -59,16 +65,32 @@ abstract class AbstractShopware6ProductCustomFieldSetMapper implements Shopware6
         return $this->getType() === $type;
     }
 
-    /**
-     * @param mixed $calculateValue
-     *
-     * @return string|array
-     */
-    abstract protected function getValue(
+    protected function createPropertyGroupOptions(
         Shopware6Channel $channel,
         AbstractAttribute $attribute,
-        $calculateValue
-    );
+        ValueInterface $value
+    ): Shopware6PropertyGroupOption {
+        $name = $this->calculator->calculate($attribute, $value, $channel->getDefaultLanguage());
+
+        $propertyGroupOption = new Shopware6PropertyGroupOption(null, $name);
+
+        foreach ($channel->getLanguages() as $language) {
+            $calculateValue = $this->calculator->calculate($attribute, $value, $language);
+            if ($calculateValue) {
+                $propertyGroupOption->addTranslations($language, 'name', $calculateValue);
+            }
+        }
+
+        return $propertyGroupOption;
+    }
+
+    abstract protected function addProperty(
+        Shopware6Product $shopware6Product,
+        AbstractAttribute $attribute,
+        AbstractProduct $product,
+        Shopware6Channel $channel,
+        ?Language $language = null
+    ): Shopware6Product;
 
     private function attributeMap(
         Shopware6Product $shopware6Product,
@@ -85,19 +107,7 @@ abstract class AbstractShopware6ProductCustomFieldSetMapper implements Shopware6
         }
 
         if ($this->isSupported($attribute->getType())) {
-            $value = $product->getAttribute($attribute->getCode());
-            $calculateValue = $this->calculator->calculate(
-                $attribute,
-                $value,
-                $language ?: $channel->getDefaultLanguage()
-            );
-
-            if ($calculateValue) {
-                $shopware6Product->addCustomField(
-                    $attribute->getCode()->getValue(),
-                    $this->getValue($channel, $attribute, $calculateValue)
-                );
-            }
+            $this->addProperty($shopware6Product, $attribute, $product, $channel, $language);
         }
 
         return $shopware6Product;
