+++ module/exporter-shopware-6/src/Infrastructure/Mapper/Product/AbstractVariantOptionMapper.php
@@ -13,13 +13,13 @@ use Ergonode\Attribute\Domain\Repository\OptionRepositoryInterface;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
 use Ergonode\ExporterShopware6\Domain\Repository\PropertyGroupOptionsRepositoryInterface;
 use Ergonode\ExporterShopware6\Infrastructure\Calculator\AttributeTranslationInheritanceCalculator;
-use Ergonode\ExporterShopware6\Infrastructure\Mapper\Shopware6ProductMapperInterface;
+use Ergonode\ExporterShopware6\Infrastructure\Mapper\ProductMapperInterface;
 use Ergonode\Product\Domain\Entity\AbstractProduct;
 use Ergonode\SharedKernel\Domain\Aggregate\AttributeId;
 use Ergonode\SharedKernel\Domain\AggregateId;
 use Webmozart\Assert\Assert;
 
-abstract class AbstractShopware6VariantOptionMapper implements Shopware6ProductMapperInterface
+abstract class AbstractVariantOptionMapper implements ProductMapperInterface
 {
     protected AttributeRepositoryInterface $attributeRepository;
