+++ module/exporter-shopware-6/src/Infrastructure/Mapper/Product/CustomField/ProductCustomFieldSetNumericMapper.php
@@ -9,18 +9,18 @@ declare(strict_types=1);
 namespace Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\CustomField;
 
 use Ergonode\Attribute\Domain\Entity\AbstractAttribute;
-use Ergonode\Attribute\Domain\Entity\Attribute\AbstractTextareaAttribute;
+use Ergonode\Attribute\Domain\Entity\Attribute\AbstractNumericAttribute;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
-use Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\AbstractShopware6ProductCustomFieldSetMapper;
+use Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\AbstractProductCustomFieldSetMapper;
 
-class Shopware6ProductCustomFieldSetTextAreaMapper extends AbstractShopware6ProductCustomFieldSetMapper
+class ProductCustomFieldSetNumericMapper extends AbstractProductCustomFieldSetMapper
 {
     /**
      * {@inheritDoc}
      */
     public function getType(): string
     {
-        return AbstractTextareaAttribute::TYPE;
+        return AbstractNumericAttribute::TYPE;
     }
 
     /**
