+++ module/exporter-shopware-6/src/Infrastructure/Mapper/Product/Decorator/NewProductPriceMapperDecorator.php
@@ -14,16 +14,16 @@ use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
 use Ergonode\ExporterShopware6\Infrastructure\Exception\Mapper\Shopware6ExporterNoMapperException;
 use Ergonode\ExporterShopware6\Infrastructure\Exception\Mapper\Shopware6ExporterNumericAttributeException;
 use Ergonode\ExporterShopware6\Infrastructure\Exception\Mapper\Shopware6ExporterProductAttributeException;
-use Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\Shopware6ProductPriceMapper;
-use Ergonode\ExporterShopware6\Infrastructure\Mapper\Shopware6ProductMapperInterface;
+use Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\ProductPriceMapper;
+use Ergonode\ExporterShopware6\Infrastructure\Mapper\ProductMapperInterface;
 use Ergonode\ExporterShopware6\Infrastructure\Model\Shopware6Product;
 use Ergonode\Product\Domain\Entity\AbstractProduct;
 
-class Shopware6NewProductPriceMapperDecorator implements Shopware6ProductMapperInterface
+class NewProductPriceMapperDecorator implements ProductMapperInterface
 {
-    private Shopware6ProductPriceMapper  $productPriceMapper;
+    private ProductPriceMapper  $productPriceMapper;
 
-    public function __construct(Shopware6ProductPriceMapper $productPriceMapper)
+    public function __construct(ProductPriceMapper $productPriceMapper)
     {
         $this->productPriceMapper = $productPriceMapper;
     }
