+++ module/exporter-shopware-6/src/Infrastructure/Mapper/Product/Decorator/NewProductStockMapperDecorator.php
@@ -1,25 +1,30 @@
 <?php
-/**
+/*
  * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
  * See LICENSE.txt for license details.
  */
 
 declare(strict_types=1);
 
-namespace Ergonode\ExporterShopware6\Infrastructure\Mapper\Product;
+namespace Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\Decorator;
 
 use Ergonode\Core\Domain\ValueObject\Language;
 use Ergonode\Exporter\Domain\Entity\Export;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
-use Ergonode\ExporterShopware6\Infrastructure\Mapper\Shopware6ProductMapperInterface;
+use Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\ProductStockMapper;
+use Ergonode\ExporterShopware6\Infrastructure\Mapper\ProductMapperInterface;
 use Ergonode\ExporterShopware6\Infrastructure\Model\Shopware6Product;
 use Ergonode\Product\Domain\Entity\AbstractProduct;
 
-class Shopware6ProductSkuMapper implements Shopware6ProductMapperInterface
+class NewProductStockMapperDecorator implements ProductMapperInterface
 {
-    /**
-     * {@inheritDoc}
-     */
+    private ProductStockMapper  $productStockMapper;
+
+    public function __construct(ProductStockMapper $productStockMapper)
+    {
+        $this->productStockMapper = $productStockMapper;
+    }
+
     public function map(
         Shopware6Channel $channel,
         Export $export,
@@ -27,8 +32,9 @@ class Shopware6ProductSkuMapper implements Shopware6ProductMapperInterface
         AbstractProduct $product,
         ?Language $language = null
     ): Shopware6Product {
-
-        $shopware6Product->setSku($product->getSku()->getValue());
+        if ($shopware6Product->isNew()) {
+            return $this->productStockMapper->map($channel, $export, $shopware6Product, $product, $language);
+        }
 
         return $shopware6Product;
     }
