+++ module/exporter-shopware-6/src/Infrastructure/Mapper/Product/ProductCoverMapper.php
@@ -1,20 +1,21 @@
 <?php
-/**
+/*
  * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
  * See LICENSE.txt for license details.
  */
 
 declare(strict_types=1);
 
-namespace Ergonode\ExporterShopware6\Infrastructure\Mapper;
+namespace Ergonode\ExporterShopware6\Infrastructure\Mapper\Product;
 
 use Ergonode\Core\Domain\ValueObject\Language;
 use Ergonode\Exporter\Domain\Entity\Export;
+use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
+use Ergonode\ExporterShopware6\Infrastructure\Mapper\ProductMapperInterface;
 use Ergonode\ExporterShopware6\Infrastructure\Model\Shopware6Product;
 use Ergonode\Product\Domain\Entity\AbstractProduct;
-use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
 
-interface Shopware6ProductMapperInterface
+class ProductCoverMapper implements ProductMapperInterface
 {
     public function map(
         Shopware6Channel $channel,
@@ -22,5 +23,13 @@ interface Shopware6ProductMapperInterface
         Shopware6Product $shopware6Product,
         AbstractProduct $product,
         ?Language $language = null
-    ): Shopware6Product;
+    ): Shopware6Product {
+        $media = $shopware6Product->getMedia();
+        if (!empty($media)) {
+            $coverMedia = reset($media);
+            $shopware6Product->setCoverId($coverMedia->getId());
+        }
+
+        return $shopware6Product;
+    }
 }
