+++ module/exporter-shopware-6/src/Infrastructure/Mapper/Product/ProductDescriptionMapper.php
@@ -13,12 +13,12 @@ use Ergonode\Core\Domain\ValueObject\Language;
 use Ergonode\Exporter\Domain\Entity\Export;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
 use Ergonode\ExporterShopware6\Infrastructure\Calculator\AttributeTranslationInheritanceCalculator;
-use Ergonode\ExporterShopware6\Infrastructure\Mapper\Shopware6ProductMapperInterface;
+use Ergonode\ExporterShopware6\Infrastructure\Mapper\ProductMapperInterface;
 use Ergonode\ExporterShopware6\Infrastructure\Model\Shopware6Product;
 use Ergonode\Product\Domain\Entity\AbstractProduct;
 use Webmozart\Assert\Assert;
 
-class Shopware6ProductActiveMapper implements Shopware6ProductMapperInterface
+class ProductDescriptionMapper implements ProductMapperInterface
 {
     private AttributeRepositoryInterface $repository;
 
@@ -42,25 +42,22 @@ class Shopware6ProductActiveMapper implements Shopware6ProductMapperInterface
         AbstractProduct $product,
         ?Language $language = null
     ): Shopware6Product {
-        $active = false;
 
-        $attribute = $this->repository->load($channel->getAttributeProductActive());
+        if (null === $channel->getAttributeProductDescription()) {
+            return $shopware6Product;
+        }
+        $attribute = $this->repository->load($channel->getAttributeProductDescription());
+
         Assert::notNull($attribute);
 
         if (false === $product->hasAttribute($attribute->getCode())) {
-            $shopware6Product->setActive($active);
-
             return $shopware6Product;
         }
 
         $value = $product->getAttribute($attribute->getCode());
-        $calculateValue = $this->calculator->calculate($attribute, $value, $channel->getDefaultLanguage());
 
-        if ($calculateValue > 0) {
-            $active = true;
-        }
-        $shopware6Product->setActive($active);
-
+        $name = $this->calculator->calculate($attribute, $value, $language ?: $channel->getDefaultLanguage());
+        $shopware6Product->setDescription($name);
 
         return $shopware6Product;
     }
