+++ module/exporter-shopware-6/src/Infrastructure/Mapper/Product/ProductSEOKeywordsMapper.php
@@ -13,12 +13,12 @@ use Ergonode\Core\Domain\ValueObject\Language;
 use Ergonode\Exporter\Domain\Entity\Export;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
 use Ergonode\ExporterShopware6\Infrastructure\Calculator\AttributeTranslationInheritanceCalculator;
-use Ergonode\ExporterShopware6\Infrastructure\Mapper\Shopware6ProductMapperInterface;
+use Ergonode\ExporterShopware6\Infrastructure\Mapper\ProductMapperInterface;
 use Ergonode\ExporterShopware6\Infrastructure\Model\Shopware6Product;
 use Ergonode\Product\Domain\Entity\AbstractProduct;
 use Webmozart\Assert\Assert;
 
-class ProductSEOKeywordsMapper implements Shopware6ProductMapperInterface
+class ProductSEOKeywordsMapper implements ProductMapperInterface
 {
     private AttributeRepositoryInterface $repository;
