+++ module/exporter-shopware-6/src/Infrastructure/Mapper/Product/ProductStockMapper.php
@@ -13,12 +13,12 @@ use Ergonode\Core\Domain\ValueObject\Language;
 use Ergonode\Exporter\Domain\Entity\Export;
 use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
 use Ergonode\ExporterShopware6\Infrastructure\Calculator\AttributeTranslationInheritanceCalculator;
-use Ergonode\ExporterShopware6\Infrastructure\Mapper\Shopware6ProductMapperInterface;
+use Ergonode\ExporterShopware6\Infrastructure\Mapper\ProductMapperInterface;
 use Ergonode\ExporterShopware6\Infrastructure\Model\Shopware6Product;
 use Ergonode\Product\Domain\Entity\AbstractProduct;
 use Webmozart\Assert\Assert;
 
-class Shopware6ProductDescriptionMapper implements Shopware6ProductMapperInterface
+class ProductStockMapper implements ProductMapperInterface
 {
     private AttributeRepositoryInterface $repository;
 
@@ -42,22 +42,22 @@ class Shopware6ProductDescriptionMapper implements Shopware6ProductMapperInterfa
         AbstractProduct $product,
         ?Language $language = null
     ): Shopware6Product {
-
-        if (null === $channel->getAttributeProductDescription()) {
-            return $shopware6Product;
-        }
-        $attribute = $this->repository->load($channel->getAttributeProductDescription());
-
+        $attribute = $this->repository->load($channel->getAttributeProductStock());
         Assert::notNull($attribute);
-
         if (false === $product->hasAttribute($attribute->getCode())) {
+            if ($shopware6Product->isNew()) {
+                $shopware6Product->setStock(0);
+            }
+
             return $shopware6Product;
         }
 
         $value = $product->getAttribute($attribute->getCode());
+        $calculateValue = $this->calculator->calculate($attribute, $value, $channel->getDefaultLanguage());
+        if (is_numeric($calculateValue)) {
+            $shopware6Product->setStock((int) $calculateValue);
+        }
 
-        $name = $this->calculator->calculate($attribute, $value, $language ?: $channel->getDefaultLanguage());
-        $shopware6Product->setDescription($name);
 
         return $shopware6Product;
     }
