+++ module/exporter-shopware-6/src/Infrastructure/Mapper/Product/ProductVariantMapper.php
@@ -24,7 +24,7 @@ use Ergonode\Product\Domain\Entity\VariableProduct;
 use Ergonode\SharedKernel\Domain\Aggregate\AttributeId;
 use Ergonode\SharedKernel\Domain\AggregateId;
 
-class Shopware6ProductVariantMapper extends AbstractShopware6VariantOptionMapper
+class ProductVariantMapper extends AbstractVariantOptionMapper
 {
     private PropertyGroupRepositoryInterface $propertyGroupRepository;
     private OptionQueryInterface  $optionQuery;
