+++ module/exporter-shopware-6/src/Infrastructure/Mapper/Product/PropertyGroup/ProductGroupMultiSelectMapper.php
@@ -8,16 +8,16 @@ declare(strict_types=1);
 
 namespace Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\PropertyGroup;
 
-use Ergonode\Attribute\Domain\Entity\Attribute\SelectAttribute;
+use Ergonode\Attribute\Domain\Entity\Attribute\MultiSelectAttribute;
 use Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\AbstractShopware6ProductPropertyGroupOptionMapper;
 
-class Shopware6ProductGroupSelectMapper extends AbstractShopware6ProductPropertyGroupOptionMapper
+class ProductGroupMultiSelectMapper extends AbstractShopware6ProductPropertyGroupOptionMapper
 {
     /**
      * {@inheritDoc}
      */
     public function getType(): string
     {
-        return SelectAttribute::TYPE;
+        return MultiSelectAttribute::TYPE;
     }
 }
