+++ module/exporter-shopware-6/src/Infrastructure/Mapper/Product/PropertyGroup/ProductGroupSelectMapper.php
@@ -8,16 +8,16 @@ declare(strict_types=1);
 
 namespace Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\PropertyGroup;
 
-use Ergonode\Attribute\Domain\Entity\Attribute\MultiSelectAttribute;
+use Ergonode\Attribute\Domain\Entity\Attribute\SelectAttribute;
 use Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\AbstractShopware6ProductPropertyGroupOptionMapper;
 
-class Shopware6ProductGroupMultiSelectMapper extends AbstractShopware6ProductPropertyGroupOptionMapper
+class ProductGroupSelectMapper extends AbstractShopware6ProductPropertyGroupOptionMapper
 {
     /**
      * {@inheritDoc}
      */
     public function getType(): string
     {
-        return MultiSelectAttribute::TYPE;
+        return SelectAttribute::TYPE;
     }
 }
