+++ module/exporter-shopware-6/src/Infrastructure/Mapper/ProductMapperInterface.php
@@ -1,0 +1,26 @@
+<?php
+/**
+ * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
+ * See LICENSE.txt for license details.
+ */
+
+declare(strict_types=1);
+
+namespace Ergonode\ExporterShopware6\Infrastructure\Mapper;
+
+use Ergonode\Core\Domain\ValueObject\Language;
+use Ergonode\Exporter\Domain\Entity\Export;
+use Ergonode\ExporterShopware6\Infrastructure\Model\Shopware6Product;
+use Ergonode\Product\Domain\Entity\AbstractProduct;
+use Ergonode\ExporterShopware6\Domain\Entity\Shopware6Channel;
+
+interface ProductMapperInterface
+{
+    public function map(
+        Shopware6Channel $channel,
+        Export $export,
+        Shopware6Product $shopware6Product,
+        AbstractProduct $product,
+        ?Language $language = null
+    ): Shopware6Product;
+}
