+++ module/exporter-shopware-6/src/Resources/config/services.yml
@@ -89,27 +89,27 @@ services:
 
     Ergonode\ExporterShopware6\Infrastructure\Builder\Shopware6ProductBuilder:
         arguments:
-            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\Shopware6ProductSkuMapper'
-            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\Shopware6ProductNameMapper'
-            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\Decorator\Shopware6NewProductPriceMapperDecorator'
-            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\Shopware6ProductCategoryMapper'
-            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\Shopware6ProductActiveMapper'
-            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\Decorator\Shopware6NewProductStockMapperDecorator'
-            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\Shopware6ProductDescriptionMapper'
-            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\Shopware6ProductGalleryMapper'
-            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\PropertyGroup\Shopware6ProductGroupSelectMapper'
-            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\PropertyGroup\Shopware6ProductGroupMultiSelectMapper'
-            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\CustomField\Shopware6ProductCustomFieldSetTextMapper'
-            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\CustomField\Shopware6ProductCustomFieldSetTextAreaMapper'
-            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\CustomField\Shopware6ProductCustomFieldSetNumericMapper'
-            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\CustomField\Shopware6ProductCustomFieldSetDateTimeMapper'
-            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\CustomField\Shopware6ProductCustomFieldSetSelectMapper'
-            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\CustomField\Shopware6ProductCustomFieldSetMultiSelectMapper'
-            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\CustomField\Shopware6ProductCustomFieldSetMultimediaMapper'
-            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\CustomField\Shopware6ProductCustomFieldSetUnitMapper'
-            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\Shopware6ProductVariantMapper'
-            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\Shopware6ProductVariantChildMapper'
-            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\Shopware6ProductCoverMapper'
+            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\ProductSkuMapper'
+            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\ProductNameMapper'
+            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\Decorator\NewProductPriceMapperDecorator'
+            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\ProductCategoryMapper'
+            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\ProductActiveMapper'
+            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\Decorator\NewProductStockMapperDecorator'
+            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\ProductDescriptionMapper'
+            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\ProductGalleryMapper'
+            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\PropertyGroup\ProductGroupSelectMapper'
+            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\PropertyGroup\ProductGroupMultiSelectMapper'
+            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\CustomField\ProductCustomFieldSetTextMapper'
+            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\CustomField\ProductCustomFieldSetTextAreaMapper'
+            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\CustomField\ProductCustomFieldSetNumericMapper'
+            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\CustomField\ProductCustomFieldSetDateTimeMapper'
+            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\CustomField\ProductCustomFieldSetSelectMapper'
+            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\CustomField\ProductCustomFieldSetMultiSelectMapper'
+            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\CustomField\ProductCustomFieldSetMultimediaMapper'
+            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\CustomField\ProductCustomFieldSetUnitMapper'
+            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\ProductVariantMapper'
+            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\ProductVariantChildMapper'
+            - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\ProductCoverMapper'
             - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\ProductSEOMetaTitleMapper'
             - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\ProductSEOMetaDescriptionMapper'
             - '@Ergonode\ExporterShopware6\Infrastructure\Mapper\Product\ProductSEOKeywordsMapper'
